moved exception handling to another file and renamed the main file to flush
