+++ src/flush.py
@@ -1,34 +1,11 @@
 import sys
-import signal
 import os.path
-from types import FrameType
 from scanner import Scanner
+from exception_handler import FileNotFoundError
 """
-* possibly add running code as a library file *
-
-REPL AND FILE RUNNING SHOULD BE IN SEPARATE FILES
--> SHOULD HANDLE SIGNALS ACCORDINGLY
+-> make module for exception handling (class)
+-> 
 """
-
-# Error handlers | CHANGE THIS TO BE BETTER
-class FileNotFoundError(Exception):
-
-    def __init__(self, message):
-        self.message = message
-        super().__init__(self.message)
-
-
-# Signal handler
-def sigint_handler(signal: int, frame: FrameType | None) -> None:
-    print('Keyboard Interrupt')
-
-def EOF_handler(signal: int, frame: FrameType | None) -> None:
-    sys.exit(0)
-
-# Register Signal
-signal.signal(signal.SIGINT, sigint_handler)
-# signal.signal(signal, EOF_handler)
-
 # REPL 
 def repl() -> None:
     stmt = ""
@@ -37,8 +14,6 @@ def repl() -> None:
         if stmt == "exit":
             sys.exit(0)
         run(stmt)
-
-        print(">", stmt) # TEMPORARY
 
 def file(path: str) -> None:
     if not os.path.isfile(path):
@@ -51,11 +26,11 @@ def file(path: str) -> None:
 
     return None
 
+# For now it scans the input outputs all of the tokens
 def run(src: str) -> None:
     scanner = Scanner(src)
-    tokens = scanner.scan()
-    for token in tokens:
-        print(token)
+    scanner.scan()
+    scanner.to_str()
 
 def main() -> None:
     if (len(sys.argv) == 1):
@@ -67,7 +42,7 @@ def main() -> None:
 
     else:
         print("Error: Invalid tool call. The tool must be in the format:")
-        print("python treewalker.py <file_name>.poo\n")
+        print("python treewalker.py <file_name>.fsh\n")
         raise SystemExit(1)
 
 if __name__ == "__main__":
